[Core] Define new Future Time Drift
# Conflicts:
#	src/chainparams.cpp
#	src/chainparams.h
#	src/kernel.cpp
#	src/kernel.h
#	src/main.cpp
#	src/wallet.cpp
#	src/wallet.h

+++ src/wallet.cpp
@@ -2982,7 +2982,13 @@ bool CWallet::CreateTransaction(CScript scriptPubKey, const CAmount& nValue, CWa
 }
 
 // ppcoin: create coin stake transaction
-bool CWallet::CreateCoinStake(const CKeyStore& keystore, unsigned int nBits, int64_t nSearchInterval, CMutableTransaction& txNew, unsigned int& nTxNewTime)
+bool CWallet::CreateCoinStake(
+        const CKeyStore& keystore,
+        unsigned int nBits,
+        int64_t nSearchInterval,
+        CMutableTransaction& txNew,
+        int64_t& nTxNewTime
+        )
 {
     // The following split & combine thresholds are important to security
     // Should not be adjusted if you don't understand the consequences
